Se termina de implementar la tabla, Funciona correctamente del, help, add
