+++ proyectos/02/ejemploTabla.txt
@@ -1,6 +1,6 @@
-fdsafds000	000	700	0000	000
-vfnytjyu00	700	140	0000	000
-vfdsdsf000	140	210	0000	000
+0000000000	000	000	0000	000
+0000000000	000	000	0000	000
+0000000000	000	000	0000	000
 0000000000	000	000	0000	000
 0000000000	000	000	0000	000
 0000000000	000	000	0000	000
