--- proyectos/02/tabla/readme.txt
@@ -1,12 +1,0 @@
-Esta es mi propuesta de implementación de la tabla, por el momento solo puede
-agregar la dirección inicial y final y nombre del archivo, falta agregarlo
-dentro del loop para que lo repita,
-
-Las entradas al correr el programa son los nombres de los archivos, el cual
-generara la ejemploTabla.txt hice 3 entradas como ejemplo.
-
-Simplificaria ya que si deseamos borrar un archivo iriamos a su nombre y
-pondriamos en cero todo, para editar un archivo, la última columna de la tabla
-indicara si se modifico el archivo, en caso de que haya sido modificado
-buscaria en toda la tabla viendo el pin con el cual coincide para enlazar los
-archivos, Falta agregar como almacenará los datos virtualmente
