--- proyectos/02/tabla/readme.txt~
@@ -1,6 +1,0 @@
-Esta es mi propuesta de implementación de la tabla, por el momento solo puede
-agregar la dirección inicial y final y nombre del archivo, falta agregarlo
-dentro del loop para que lo repita,
-
-Las entradas al correr el programa son los nombres de los archivos, el cual
-generara la ejemploTabla.txt
